Update device local control configuration.
Now gateway auto discovery is done right on app start. If the gateway
responds through socket, with a uuid, localControl is set true for that.
For those gateways support local control, request is made to get token
for local control. Also on connectivity change, the gateways are auto
discovered again to update the new IP address.