Provide property 'onPress' to the SwipeRow visible content.
'ListItem', which is not a touchable but still onPress is passed inorder
to fix issue- cannot control device from device list in accessibility
mode(ios). Issue is when the visible content(ListItem) of SwipeRow
does not have a 'onPress' property it is placed inside a touchable
component, which catches the focus and prevent children(device button)
from being accessible.

See  #389.