Prevent animated 'X' value from going off-track
This happens on changing schedule week days by sliding the poster.
This commit is only a prevention from value going off the limit, and
this has a down side

1. Beyond the distance of - next day becoming today/prev day becoming
today, the day will not move even if user keep sliding.
2. After sliding beyond the above mentioned threshhold, when user slide
in reverse direction, it takes user to slide back all the neutral place again
for the reverse-animation to happen.

Fixes #406.