Remove RCLCM 'componentWillReceiveProps' - PreLogin screens
Now the logic is handled using combination of 'getDerivedStateFromProps'
and 'componentDidUpdate'. Navigation on prop change also set state han-
 dled inside 'getDerivedStateFromProps'. Async call handled at 'compone-
ntDidUpdate'.

Other Change:
- Since button labels are dependent on loading property/state, now the
async API call's catch block is handled to reset the state. It is not req-
uired to handle then block(infact should not be), as it cause navigation
or unmount of the screen.

See #305.