Replace RCLCM 'componentWillReceiveProps' with 'getDerivedStateFrom- Props' - DimmerProgressBar
One major change here is method 'animate' is moved outside the class and
'animationValue' is passed as an argument. The reason for doing so is, the
method need to be accessed from the static method 'getDerivedStateFro-
mProps'. A different alternative would be to introduce a new state and
update it where animation is done now, and inside 'DidUpdate' do the ani-
mation instead, this would cost performance due to frequent set state
while dimmer progress. So the current implementation looks better.

See #305.