Correct those naive mistakes
!- Use 'navigate' instead of 'push', now only grasping the actual
behavioural difference, switching back to 'navigate' because we
do not want the screens to stack up, instead if already visited p-
reviously it is okay to navigate to it. difference can be identified
in android, keep switching between pre-login screen then use
android back button to traverse back.
!- Use camel case while naming reducers, renaming the ‘App’
reducer to ‘app’, updates all instances where it is used.