Update 'WidgetModule' exposed methods
Method 'configureWidgetData' is renamed as 'configureWidgetAuthData'.
Also now it takes 5 arguments all of string type. And this is meant for
passing auth related(token and keys) data, down to the widget/native
side.

Also adding a new method 'configureWidgetSessionData', As of now takes
single parameter, that is 'session id'.  The session data is passed down
to widget side, at two occasions,
1) If available, at will mount of post-login root view, along with
auth data.
2) If available, right after websocket session authentication.
This is to make sure it is always passed.