Refactor Android widget activities and utilities
This commit removes unnecessary data stored/passed to native/widget
side, this includes,
-  user name
- password
- token type
- grant type
- scope
- ttl
Also removing all instance of reading the auth data and session data
from file in native file system. Instead it is passed with the help of
NativeModule 'WidgetModule' and is stored in shared preference.
And thereby must be available at all Activities.