Add method 'disableAllWidgets' to native WidgetModule
Expects one argument of string type, which is the message to be
displayed on disabled widgets. This can be used at cases like
log out from app, and since widgets cannot be removed from home
screen programatically, using this method it can be disabled and
any helpful message can be shown.