Set a timeout of 3secs for local device control and local end point test
Sometimes when the local control address is not reachable or
something, there is a chance for the API request to not resolve/
reject for quite a long time. By setting timeout it will make sure
request is rejected and catch block is called after 3secs.

Upon rejection in the case of device control, request will be completed
through cloud, and 'supportLocal' property will be set as false. And in
the case of end point test, local control address is reset/cleared.