Remove axios implicit timeout, set explicitly instead, also cancel request on timeout - Local Control/Test
implicit timeout works under normal condition, but when the ip
address is not reachable, the axios does not adhere to the timeout,
this is frequent case in local control/test.

To overcome time, implicit timeout is removed, and timeout is handled
explicitly, also upon timeout the corresponding request is cancelled
explicitly.

Also this commit fixes a bug, and make sure each different timeouts
and timeinterval variables are unique. so that clearing each does clear
the expected one alone.