Fix issue backtracing sourcemap always throw null:null
Issue is of two parts, first is fixed by this commit, second is
yet to be figured out.

First issue is at some case the 'line' and 'file' value are not integers
(could be dependent on the editor in which the 'stack trace' is
formatted), so this cannot be an issue for all. Now this commit makes
sure it is always integer.

The second part, the backtraced file/position/line are found to be
wrong. This can be related to the RN issue, that wrong sourcemap
being created.

See #524.