Make sure 'this.device' is never undefined - Schedule/Edit screen
The chances are rare, for a schedule to be present and the device
object to go undefined in schedule Edit screen. Yet, as there is crash
reported in v3.7.7(iOS)(TypeError: undefined is not an object
(evaluating 'a.device.name')), this commit makes sure it is an empty
object and accessing properties on it will not crash.