Update UDP socket handle during Tellstick local discovery
This commit makes sure 'send' is not called on any socket object,
on which 'close' is called, and will soon be closed.
'bind'/'close'/'send' are all asynchronous and there is a error
'send failed: no client found with id ##' to happen. This is a
non-fatal and crash was due to mis-handling of 'reportException'
method.