Always use same device unique id while push token register
In iOS atleast during OS update or some other case the values
returned by 'DeviceInfo.getUniqueID()' does seem to change.
This results in multiple record entries in server and issues due
to that. To solve this even the 'IDFV' change, we use previosuly
registered ID to make the request.

This will only be effective on user's who install the app fresh,
and for those who has already installed during the very next
OS update, if 'IDFV' change from that of previous, that will
not be handled, because then only deviceid is stored in redux
store for the first time. After that it will not happen.

Fixes #437.