Disable local control on iOS 9 and lower
In iOS library 'react-native-rsa-native' can only be required/used in 10
or above. In order to get the app running in iOS 9, only alternative is
to disable local control, untill a better solution is found. Now that the
lib is weak linked, compile time issue will not appear, and not this
conditional require only in supported platforms will get the app
running.

Changelog: App supported in iOS 9. Local control only in iOS 10 above.