Change animation while opening/closing wizard nav screens
This new transition animation is applicable to all scree/sub navigators
wired in root navigator. Now the screens appear from bottom and closes
through bottom.

Some changes that were made to achieve this pleasantly:
- The HeaderPoster of both Device/Sensor details wizard is now
render within the Tab bar component. It is important to move it out
of root navigator.(Unconventional)
- The Main Tab Header has so far been rendered as a separate compo-
nent out side the Navigator, now in iOS it is always rendered inside
Stack as a Stack header component, In android also the same except
when in Landscape mode

See #554.