Fix sensor value text size in Android
Issue is that when user has font size increased in device's accessib-
ility settings, the sensor value text alone does not seem to obey the
allowFontScaling={false} property. This happens only when
FormattedNumber component is rendered inside another 'Text'
component, and only in Android.

As a fix, this commit refactores the component 'FormattedNumber' and
'GenericSensor' to render the sensor unit as suffix also accept custom
style on the suffix text.

Fixes #568.