Switch to fabric provided by react-native-firebase
It is crashlytics for which the fabric is used. Since the complete
fabric ecosystem is switching to firebase, we are also aiming to
avail crashlytics from the firebase module we already use. The
RN firebase module might still have to migrate internally, but
anyway we can expect lesser changes if switched to a firebase
module.

This commit, majorly unlinks and uninstalls the module 'react-
native-fabric'. Now all methods(almost) are used from 'react-
native-firebase'

See #600.