Update schedule navigator!
- Now passing a new property 'ScreenName' which is the screen name
of each route, down to the Navigator Container Component.
- renderScheduleScreen method is called for every screen in a stack
navigator when route change, this will result in Navigator Container
Component to render every single time. This causes some issue on
showing modal from the last screen.

Issue is modal background is going black. Now using 'ScreenName'
container component for those screens which are not in focus is
prevented from updating, and the issue is resolved.

- Also this commit solves issue - Scheduler editing help not working

See #625.