Handle device history details modal using local state
This is change is to not store the device history data which is an object
type in redux store as Modal reducer's data attribute. And there by can
make sure, Modal reducer's data attribute is always of string type.

There is a crash reported in both iOS and Android, which on backtracing
points to RN code/file, which kind of give some clues that it is due to
render returning some object type(similar to device history data type).