Update App.js file - renders root dialoguebox
The root file renders a DialogueBox, which is now controlled using a
callback method passed down as a property.

Later on we could remove dialogueboxes used inside sub navigators and
sub components with this root dialogue. The control method will be one
challenge, as it contains some props which are methods(onPress), so
no a good idea to store method in redux store(else could use the
existing thunk action[showModal] as control method)