Fix issue with ButtonLoadingIndicator, reported in Android P
Issue seem to be with animating the opacity of vector-icon, with
background color being transparent or not provided. One solution
would be to set background as same as the button beneath, but
that would be hectic as the button color can vary alot and will be
a lot of changes.

This commit replaces round icon with rounded view.