Update device inclusion/exclusion process
Major change is that, once the socket 'onmessage' listener is
set,it will be the only instance working then on. It is not possible to
have multiple socket listener's of same gateway active at the same
time.

Due to this it is important to handle all normal messages
(processWebsocketMessage). This commit does that and makes
sure, socket works without any issues post - inclusion/exclusion.

Another change is this commit makes sure the timeouts are clear
properly.