Show device action icons based on 'deviceType'
This commit implements 'deviceType' based icons for some devices.
Devices like bell, navigational are yet to be updated.

Unlike app the  widget implementation of icons are different that
is why some are left out, to make it global, need to make some
changes in how telldus icons are rendered inside widget.

See #735.