Handle firebase crashlytics reporting gracefully
During development there in an adjustment to be made in
Android for firebase to not show error, that crashlytics is
not enabled before calling singleton.

Issue is because, crashlytics is enabled only in release mode
where as reporting and other things are done at different
instances, regardless of dev/release mode. iOS internally
seem to check if crashlytics is enabled before reporting, or
atleast we have no issues there.

As a fix now all crashlytics actions are moved to 'Lib/Analytics'
and is made sure it is not in DEV mode, before calling.