Trigger refresh android widgets method on app start
This method will call the native refresh methods with latest devices
and sensors list of id's. And that is used to check along side added
widgets, those widgets which are present in the latest API list is
retained as such, and others are nullified(id set to -1), which on
next update will be disabled.

See #515.