Set up sensor widget API polling framework
This commit implements the framework for sensor widget API poling.
That means, for each sensor a pair of handler and runnable is created
and maintained in a static 'Map', also on following cases the respective
handler's runnable callback is removed and the pair is removed from
the Map.
- Removed when widget is deleted.
- Removed when widget's deviceId is nullified(set to -1)(few cases).
- Removed when user has switched account/logged out.

As mentioned this is the framework alone, the actual API poling request
has to be implemented yet inside the runnable.

See #756.