Remove class 'AccessTokenService' - Android
We do not need refresh access token to run as a service. App updates
the widget side pref manager when ever token is refreshed by the app.
We do not have to manually calculate expiry date, instead only need to
listen for corresponding error at API request and refresh only then.
It is already done, so removing the previous code.