Update class 'DevicesAPI'
This commit introduces some enhancements and changes to the
previous 'setDeviceState' and 'getDeviceInfo' methods.

Now after setDeviceState success the deviceInfo check to confirm
the current state is done every 2secs for the next 10secs, previously
it was 1sec and that can cause multiple requests to be fired, before
first one responds and call result in calling callbacks multiple times.
Apart from that another measure to prevent the same issue is to ch-
eck handler/runnable was cleared before proceeding any action after
deviceInfo check success.

Another change is methodRequested attribute is not to null right after
API response/error, the attribute is required to decide success/error
state on the UI.