Update Rn from 0.59.5 to 0.60.4
There is a lot of changes with this upgrade. To be compatible with
latest RN, some of the dependencies and dev dependencies are also
upgraded.

Most of the dependencies are now auto-linked, there are still some
which can be found inside 'react-native.config.js' file.

Another change is from now on Android used vector icons via gradle
which is recommended, rather than manual copying to assets.

See #859.