Update component 'DeviceSettings'
Do blur the input settings when keyboard is hidden.

In Android, when keyboard is hidden by user pressing the
device's back button, the input is still not blurred. This will
result in the user to press again one on the screen to blur,
only after which other press/touch will happen.