Update action set app layout
Previously keyboard event listeners were used to
check if it is shown and if so app layout is not updated
in store.

Issues with above method is in android(atleast) when
orientation is landscape, keyboard events are fired(no idea
why). This cause issue updating main tab bar in android
with orientation change. Another issue is with keyboard
open if user changes device orientation layout is not
updated and styling issues occur.

This commit check for cached width and height and updates
the store only if both has changed.