Supply initial device action while selecting  device
When a device is selected for perform some action on
fence entry/leave supply current state as initial action.
If user has not explicitly selected an ection this will be
used.

The Thermostat and RGB supported devices are exclusion
from the above, for those user has to explicitly choose some
action.

See #43.