Initial commit - support for in-app update
This commit adds support to check for
in-app update on app start(not resume),
regardless of log in status. Both in iOS and
Android.

If any update is available, a dialogue is
shown asking the user to authorise update
If allowed, in iOS user is taken to app store
and in Android, the flow afterwards is yet to
decide and implement.

See #1023.