Use right pair of accesstoken and user id at widget side
With the multiple accounts feature, and since accesstoken
is updated in redux before user profile, once first a wrong pair
of access token and user id was cached in widget side, which
is then overridden when the new user profile is fetched and
updated at redux.

This commit fixes that issue and makes sure even at the
first time, right pair is used.