Update action caching dimmer initial state
This is done right before a dimmer value is
controlled/changed. The initial value is cached
in a different store object to reset later if the
action fails.

The change is not to cache this value if an
action was already performed and is not
resolved yet. Doing so will result in wrong
value to be cached as initial value.