Refactor iOS 'WidgetModule'
The expose native methods are intact, but the actual
function implementations are now moved to a new class/
file - 'SharedModule', WidgetModule simply calls those.
The idea is to share 'SharedModule' across different app
targets/extensions, and not to share class 'WidgetModule'
itself, since it contains React Bridge related methods which
result in build errors as they are not available for extensions.

This means we can now use RCT promises and callback inside
'WidgetModule'.