Use Changelog screen as a navigator screen
Previosuly it was shown as a Modal. And since it is
shown on app start, asynchronously, due to a bug
in RN Modal, some native Modals does conflict
with the RN Modal and blocks the entire UI when
one is shown when another is already there.

The workaround is to wire the changelog as navigator
screen.